// Angular_Ariketak_3/jatetxeak/src/app/interfaces/jatetxea.ts
export interface Jatetxea {
    documentName:                string;
    documentDescription:         string;
    templateType:                TemplateType;
    qualityQ:                    string;
    qualityIconDescription:      QualityIconDescription;
    accesibility:                string;
    accesibilityIconDescription: AccesibilityIconDescription;
    marks:                       Marks;
    physical:                    string;
    visual:                      string;
    auditive:                    string;
    intellectual:                string;
    organic:                     string;
    qualityAssurance:            string;
    tourismEmail:                string;
    web:                         string;
    importance:                  string;
    room:                        string;
    productClub:                 string;
    visit:                       string;
    capacity:                    string;
    store:                       string;
    gastronomical:               string;
    surfing:                     string;
    postalCode:                  string;
    restorationType:             RestorationType;
    recomended:                  string;
    recomendedURLIcon:           RecomendedURLIcon;
    recomendedIconDescription:   RecomendedIconDescription;
    restaurant:                  string;
    bodega:                      string;
    michelinStar:                string;
    repsolSun:                   string;
    latitudelongitude:           string;
    latwgs84:                    string;
    lonwgs84:                    string;
    placename:                   string;
    municipality:                string;
    municipalitycode:            string;
    postalcode:                  string;
    territory:                   Territory;
    territorycode:               string;
    countrycode:                 string;
    email:                       string;
    webpage:                     string;
    friendlyUrl:                 string;
    physicalUrl:                 string;
    dataXML:                     string;
    metadataXML:                 string;
    zipFile:                     string;
    phone?:                      string;
    address?:                    string;
    country?:                    string;
}

export enum AccesibilityIconDescription {
    Empty = "",
    EzinduentzatEgokitua = "Ezinduentzat egokitua",
}

export enum Marks {
    ArabakoErrioxa = "Arabako Errioxa",
    Bilbao = "Bilbao",
    Donostia = "Donostia",
    Empty = "",
    EuskalKostaldea = "Euskal Kostaldea",
    EuskalKostaldeaBilbao = "Euskal Kostaldea,Bilbao",
    EuskalKostaldeaDonostia = "Euskal Kostaldea,Donostia",
    EuskalKostaldeaGureMendiakEtaHaranak = "Euskal Kostaldea,Gure Mendiak eta Haranak",
    GureMendiakEtaHaranak = "Gure Mendiak eta Haranak",
    GureMendiakEtaHaranakBilbao = "Gure Mendiak eta Haranak,Bilbao",
    GureMendiakEtaHaranakDonostia = "Gure Mendiak eta Haranak,Donostia",
    VitoriaGasteiz = "Vitoria-Gasteiz",
}

export enum QualityIconDescription {
    Empty = "",
    KalitateQAgiria = "Kalitate Q agiria",
}

export enum RecomendedIconDescription {
    Empty = "",
    GidakGourmetourMichelinRepsol = "Gidak: Gourmetour, Michelin, Repsol",
}

export enum RecomendedURLIcon {
    Empty = "",
    RecomendadoenguiasGIF = "recomendadoenguias.gif",
}

export enum RestorationType {
    Erretegiak = "Erretegiak",
    Jatetxeak = "Jatetxeak",
    Sagardotegiak = "Sagardotegiak",
}

export enum TemplateType {
    SukaldaritzaAlorra = "Sukaldaritza-alorra",
}

export enum Territory {
    ArabaÁlava = "Araba/Álava",
    Bizkaia = "Bizkaia",
    Gipuzkoa = "Gipuzkoa",
    GipuzkoaGipuzkoa = "Gipuzkoa Gipuzkoa",
}